Function dominion::buy

pub fn buy(c: Card) -> Result

Buy a card from the supply, returning one of three possible errors:

NotInSupply, if the card is not available in this game, or

EmptyPile, if there are no more available to buy, or

NotEnoughMoney(need, have), if the player doesn't have enough money.

On success, the appropriate supply count is decremented and a copy of the card is added to the player's discard pile.