Trait dominion::Player

pub trait Player {
    fn name(&self) -> &'static str;
    fn init(&self, kingdom: &[Card]) -> fn();

    fn library_should_discard(&self, drawn: Card) -> bool { ... }
    fn militia_discard(&self, hand: &[Card]) -> Card { ... }
    fn moat_should_block(&self, attacker: Card) -> bool { ... }
    fn spy_should_discard(&self, top_card: Card, is_self: bool) -> bool { ... }
    fn bureaucrat_use_victory(&self, victory_cards: &[Card]) -> Card { ... }
    fn thief_trash_and_keep(&self, options: &[Card]) -> (Card, bool) { ... }
}

A player definition.

The only required methods are name() and init(), but other methods may be overridden in order to gain more control over your player.

Required Methods

fn name(&self) -> &'static str

Gets the name of this player, which must be unique.

fn init(&self, kingdom: &[Card]) -> fn()

Called before the first turn is played. kingdom is a slice of the 10 cards that will be used this game, and it should return a pointer to the function that will be called once per turn.

Provided Methods

fn library_should_discard(&self, drawn: Card) -> bool

Called when an Action card is encountered as part of a Library draw. It should return true if that card should be discarded, false if it should be kept.

By default, always discards Action cards.

fn militia_discard(&self, hand: &[Card]) -> Card

Called when another player plays Militia, and is called repeatedly until you have three or fewer cards in hand. The value returned should be one of the cards in your hand. Returning a card not in your hand results in a task failure.

By default, discards the first card.

fn moat_should_block(&self, attacker: Card) -> bool

Called when another player plays an attack card while you have a Moat in hand. It should return true if you wish to block the attack, otherwise false.

By default always blocks attacks.

fn spy_should_discard(&self, top_card: Card, is_self: bool) -> bool

Called when a Spy is played, including by you. Given the value of the top card of a player's deck, this method should return true if that card should be discarded, and false if it should be returned to the top of the player's deck. The value of is_self is true if and only if you are the player being acted on.

By default, keeps victory and curse cards on top for other players and discards anything else, and does the exact opposite for yourself.

fn bureaucrat_use_victory(&self, victory_cards: &[Card]) -> Card

Called when another player plays Bureaucrat, and you have one or more Victory cards in your hand. victory_cards is a slice of all the Victory cards in your hand (and will always contain at least one), and the value returned will be placed back on top of your deck. Returning a value not in the slice results in a task failure.

By default, returns the first card.

fn thief_trash_and_keep(&self, options: &[Card]) -> (Card, bool)

Called when you play Thief and someone reveals one or more treasure cards. options contains at least one card (but no more than 2), and it should return a tuple describing how to treat the reveal. The first value is the card that should be trashed, and the second value is a boolean indicating whether or not it should be kept.

By default, always trashes the highest value treasure card and keeps it.